refactor(sync): app-level sync orchestration была вынесена в отдельный orchestrator
в AppDependencies удалено прямое управление SyncCoordinator lifetime и remote reload correlation, а эта orchestration-логика перенесена в AppSyncRuntimeOrchestrator
это сужает обязанности AppDependencies вокруг composition graph и оставляет отдельную app-level точку для bootstrap-derived sync enablement, runtime source wiring и store-scoped remote reload flow
